+++ R/group_by.R
@@ -12,7 +12,6 @@
 NULL
 
 
-
 RPolarsGroupBy = new.env(parent = emptyenv())
 
 #' @export
@@ -25,28 +24,30 @@ RPolarsGroupBy = new.env(parent = emptyenv())
 #' @export
 `[[.RPolarsGroupBy` = `$.RPolarsGroupBy`
 
-#' @title auto complete $-access into a polars object
-#' @description called by the interactive R session internally
-#' @param x GroupBy
-#' @param pattern code-stump as string to auto-complete
-#' @return char vec
 #' @export
-#' @inherit .DollarNames.RPolarsDataFrame return
 #' @noRd
 .DollarNames.RPolarsGroupBy = function(x, pattern = "") {
   paste0(ls(RPolarsGroupBy, pattern = pattern), "()")
 }
 
-
 #' The internal GroupBy constructor
 #' @return The input as grouped DataFrame
 #' @noRd
 construct_group_by = function(df, groupby_input, maintain_order) {
-  if (!inherits(df, "RPolarsDataFrame")) stop("internal error: construct_group called not on DataFrame")
-  df = df$clone()
-  attr(df, "private") = list(groupby_input = unlist(groupby_input), maintain_order = maintain_order)
-  class(df) = "RPolarsGroupBy"
-  df
+  if (!inherits(df, "RPolarsDataFrame")) {
+    stop("internal error: construct_group called not on DataFrame")
+  }
+  # Make an empty object. Store everything (including data) in attributes, so
+  # that we can keep the RPolarsDataFrame class on the data but still return
+  # a RPolarsGroupBy object here.
+  out = c(" ")
+  attr(out, "private") = list(
+    dat = df$clone(),
+    groupby_input = unlist(groupby_input),
+    maintain_order = maintain_order
+  )
+  class(out) = "RPolarsGroupBy"
+  out
 }
 
 
@@ -58,13 +59,13 @@ construct_group_by = function(df, groupby_input, maintain_order) {
 #' @return self
 #' @export
 #'
-#' @examples pl$DataFrame(iris)$group_by("Species")
+#' @examples
+#' pl$DataFrame(iris)$group_by("Species")
 print.RPolarsGroupBy = function(x, ...) {
-  .pr$DataFrame$print(x)
-  cat("groups: ")
   prv = attr(x, "private")
-  cat(toString(prv$groupby_input))
-  cat("\nmaintain order: ", prv$maintain_order)
+  .pr$DataFrame$print(prv$dat)
+  cat("groups:", toString(prv$groupby_input))
+  cat("\nmaintain order:", prv$maintain_order)
   invisible(x)
 }
 
@@ -86,6 +87,7 @@ print.RPolarsGroupBy = function(x, ...) {
 #'   pl$col("bar")$mean()$alias("bar_tail_sum")
 #' )
 GroupBy_agg = function(...) {
+<<<<<<< HEAD
   if (isTRUE(attributes(self)[["is_rolling_group_by"]]) ||
       isTRUE(attributes(self)[["is_dynamic_group_by"]])) {
     class(self) = "RPolarsLazyGroupBy"
@@ -102,6 +104,15 @@ GroupBy_agg = function(...) {
     class(self) = "RPolarsGroupBy"
   }
   out
+=======
+  prv = attr(self, "private")
+  prv$dat$lazy()$group_by(
+    prv$groupby_input,
+    maintain_order = prv$maintain_order
+  )$
+    agg(...)$
+    collect(no_optimization = TRUE)
+>>>>>>> main
 }
 
 
@@ -304,6 +315,7 @@ GroupBy_null_count = function() {
 #'
 #' gb$ungroup()
 GroupBy_ungroup = function() {
+<<<<<<< HEAD
   if (isTRUE(attributes(self)[["is_rolling_group_by"]]) ||
       isTRUE(attributes(self)[["is_dynamic_group_by"]])) {
     class(self) = "RPolarsLazyGroupBy"
@@ -314,4 +326,8 @@ GroupBy_ungroup = function() {
     attr(self, "private") = NULL
   }
   self
+=======
+  prv = attr(self, "private")
+  prv$dat
+>>>>>>> main
 }